// packages/ghl-snapshots/src/snapshot-types.ts
export enum SnapshotType {
  LOCATIONS = 'locations',
  PIPELINES = 'pipelines',
  WORKFLOWS = 'workflows',
  CALENDARS = 'calendars',
  AI_WORKERS = 'ai_workers',
}